abstracted out action chains
