+++ py/selenium/webdriver/common/platforms.py
@@ -1,0 +1,11 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+__author__ = 'Tim Martin'
+
+ANY = 'ANY'
+WINDOWS = 'WINDOWS'
+ANDROID = 'ANDROID'
+MAC = 'MAC'
